issue-233: use terraform-plugin-mux
This allows both the legacy terraform-provider-sdk and the new
provider framework to be used concurrently.
https://developer.hashicorp.com/terraform/plugin/mux/combining-protocol-version-5-providers

Additional notes:

The DefaultFunc Schema option no longer works in the plugin Framework.  Default values
are instead  provided in the Configuration function.  In order for the plugin mux
to work properly the provider Schema must match between the sdk implementation
and the framework implementation.

See discussion for more details:
https://discuss.hashicorp.com/t/muxing-upgraded-tfsdk-and-framework-provider-with-default-provider-configuration/43945

This means we have to remove the DefaultFunc config from the SDK provider schema so that it matches
the new Framework provider schema.  Default values for token and URLs are now set in the
`configure` function instead.