update tfdocs dependency and minor CI fixes (#259)
This upgrades the terraform-docs-lib dependency which results in some
minor doc updates.

This also fixes a bug in the streaming namespace acceptance test. The streaming
namespace test is sometimes failing due to trying to create the
Pulsar namespace before the tenant is created.  This addes a dependency relation
to ensure the streaming tenant is created first.