issue-244: streaming pulsar namespace policy management (#267)
* issue-244: streaming pulsar namespace policy management

This adds basic support for managing Pulsar namespace policies.  We currently
only support a limited set of the policies for simplicity, but we can easily add
more as needed.  This also introduces a new dependency on the pulsar admin client
go library.

Because the policy configuration is complex with multiple levels and types of
attribute nesting, the Terraform plugin protocol version 6 was a better fit
for this change, so this also does some refactoring of the providers to use
protocol version 6.  This means that Terraform CLI veersions before 1.0
will no longer be supported.

* compatibility with go 1.18