STREAM-166: fix bug in CDC resource
This fixes an issue when the local CDC state is out of sync with the remote resource state
and the plugin crashes with an index out of bounds on the CDC status slice.

The issue happens when the user does the following
1. create CDC configs using Terraform
2. manually delete all CDC configs for that database via the UI
3. run terraform apply again

This also includes some other minor code cleanup.