issue-227: improve error handling for access list resource
Due to a bug in the astra go client, if the server responds with a 403 due to limited
permissions, a JSON parsing error is reported to the user instead of reporting the
actual permissions issue.  This works around that by moving the response parsing
code into the terraform provider.

This also adds a couple of nil pointer checks to prevent plugin crashes in certain
situations.