Remove streaming resources on out-of-band updates
This patch will remove sinks, tenants and topics from
the TF state if, on READ, they have been reomved outside
of the TF process. For example, if one of these resources
is removed via the GUI, and not by removing the resource
from the TF manifest, performing a --refresh-only will
remove the resource from the TF state.